refactor(compiler-plugin): extract transformPrivatePreviewToInternal feature
Pure move of `PreviewLabFirStatusTransformerExtension` into its dedicated
feature directory:

  fir/PreviewLabFirStatusTransformerExtension.kt
  → feature/transformPrivatePreviewToInternal/fir/visibilityPromotion/
       PreviewLabFirStatusTransformerExtension.kt

Updates `PreviewLabFirExtensionRegistrar` to import from the new path. Class
name keeps the `Extension` suffix per Kotlin compiler API convention
(`FirStatusTransformerExtension` continues to drive the contract).

Quality gates (all green):
- `./gradlew ktlintFormat ktlintCheck --continue` — EXIT 0
- `./gradlew :compiler-plugin:test --rerun-tasks --continue` — EXIT 0
- `./gradlew jvmTest --rerun-tasks --continue` — EXIT 0
- `./gradlew apiCheck --continue` — EXIT 0
